pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rbot"
description = 'Another bot for talking with radarr'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Eduardo Enriquez (eduzen)", email = "eduardo.a.enriquez@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["python-telegram-bot", "httpx", "rich", "python-decouple"]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
  "coverage",
  "ipdb",
  "ipython",
  "mypy",
  "pytest",
  "pytest-asyncio",
]

[project.urls]
Documentation = "https://github.com/eduzen/rbot#readme"
Issues = "https://github.com/eduzen/rbot/issues"
Source = "https://github.com/eduzen/rbot"

[tool.hatch.version]
path = "rbot/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=rbot --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "rbot/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[tool.ruffus]
ignore = ["E501"]
